core(fr): move PerformanceObserver installation to wait conditions
